requirements.txt
fastapi==0.115.11
uvicorn
gunicorn
sqlmodel
pydantic
sqlalchemy
timescaledb
python-dotenv
psycopg[binary]